requirements.txt
Cython
matplotlib
pandas
numpy
torch
pytorch-lightning
wandb
python-dotenv    